server/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<0.31
httpx>=0.28,<0.29
pillow>=10,<11
